refactor: Add `TabCell`
