+++ MyFlow/MyNavigationView/Views/Subviews/DocumentTabsCollectionViewAdaptor.swift
@@ -96,23 +96,3 @@ extension DocumentTabsCollectionViewAdaptor: UICollectionViewDropDelegate {
         }
     }
 }
-
-
-
-final class TabCell: UICollectionViewCell {
-    lazy var label = UILabel()
-    
-    override init(frame: CGRect) {
-        super.init(frame: .zero)
-        backgroundColor = .systemMint
-        
-        contentView.addSubview(label)
-        label.snp.makeConstraints { make in
-            make.centerX.centerY.equalToSuperview()
-        }
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
